backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-openai
langchain-community
chromadb
PyPDF2
openai
tiktoken
langchain-groq
sentence-transformers
langchain-huggingface
psycopg2-binary
pgvector
